InstalledCommand moved to dedicated installer utils module
