+++ tests/foreman/destructive/test_installer.py
@@ -19,7 +19,7 @@
 import pytest
 
 from robottelo.config import settings
-from robottelo.helpers import InstallerCommand
+from robottelo.utils.installer import InstallerCommand
 
 pytestmark = pytest.mark.destructive
